Some small bounds/library-include fixups.
I tried pulling & compiling this old project and despite many warnings
about deprecated GLUT features, it does work! It just took a little bit
of coaxing and fixing some of these silly mistakes to get it to play.

Tested on OSX 10.13.6, which I guess just has GLUT on it somewhere.