Clean up unused graphics.c file.
This file is unused, but it represents the code given by the course's
professor for the final stage of the assignment. It's being kept around
while I work on its successor in engine.c to make it easy to figure out
what extras I'd implemented over a decade ago when I did this work.