Fix digging, add crosshair+cube-highlight.
The calculations of direction were probably always basically right,
with the exception that they weren't converting to radians -_-' --
that likely was some kind of 4am-edit-induced error back in 2011.

Once the radians thing was fixed, basic digging worked!!! Of course,
I'd re-figured it all out along the way, so I kept the newer code.

Then, I made that into a bit of a raycast-walk style dig, to step
along the direction line until a non-empty cube is found, and to clear
just that first cube.

Then, I made it so that the raycast runs all the time (doesn't seem to
slow us down too much), and it'll draw a lines-only highlight around
the block it hits, so you can see what block you're going to interact
with, a la minecraft. This also reduced the dig if-block hugely, but
meant a (hopefully temporary) addition of some new global state for
the highlight.

And finally, I added in a crosshair by appending a brief jump into
orthographic projection and drawing some quads there.

This messed with the frustum extraction and I'm not sure why...

Previously, we were just living in a certain projection I guess, and
once I added the "switch to ortho / switch back to perspective / load
identities" code, the frustum extraction seemed to be clipping most of
what we wanted to render out. By moving the display list build to after
a few of the initial transforms we do (translation, rotation, etc), we
got clipping back to a healthy place. Clearly a lot I am doing wrong or
don't understand there still.