Overhaul player motion.
Player movement had a lot of issues, many of which are now fixed or
at least in a much better spot.

Some issues with how the code was before this commit:

a) the player's position was stored in negative co-ordinates, because
   it represented the camera's location, which was supplied as a matrix
   translation -- pushing the whole world by -10 units brings you 10
   units toward your goal, so the player (and the camera) were always
   at their -x,-y,-z co-ordinates. But everything else (like the cubes
   of the contents of the world) are stored at their _positive_ number
   locations. So a lot of multiplying by -1 and subtracting when you
   mean adding results. It's very confusing.

b) all motion updates were tied to the keyboard() function, which is
   only fired in response to keydown events at the rate the OS gives
   them to the program -- which can be especially weird due to most
   OS' default debouncing behaviours.

c) motion updates directly moved the player around in the world by a
   set amount every time they fired, making for a very chunky skippy
   feeling of hopping around -- it makes the whole game look like it
   has a low framerate

d) the player's co-ordinates are stored in the global vpx,vpy,vpz
   values, which are accessed willy-nilly all over the place, and
   *also* stored in the player's slot in the playerPositions array,
   and the two are being kept in sync throughout the game.

e) gravity is just subtraction by a fixed amount, and similarly to
   other keyboard-driven motion, looked very choppy and weird.

So, this commit does the following:

a) flips the value of the player's position to positive numbers and
   just multiplies them by -1 when it does the rendering translations

b) introduces a keyup function and a keyState array that store the
   current keyboard state, and moves motion handling from the keyboard
   function over to the update function, so it will be applied more
   or less continuously and separately from the OS' keyboard event
   rate and debouncing logic.

c) introduces pxspeed,pyspeed,pzspeed variables (which are, for now,
   new global state), and increments them by an acceleration value
   up to a max speed when motion is handled. It also decelerates the
   player using a linear friction constant applied backwards relative
   to the player's current motion.

d) drops the vpx,vpy,vpz variables entirely in favour of using the
   playerPosition array only, and removes the collisionResponse
   function which is now implemented in-line (ick) in the update
   function. The mvx,mvy,mvz variables still exist and will be
   fully removed down the road -- likely alongside a move to a
   more fulsome Player data type, or fixes to the mouse handling.

e) gravity is now also an acceleration/max-fall-speed calculation
   done during updates. There are some remaining issues here related
   to max fall speed and potentially clipping through single block
   floors at high speeds.

This overall means a bit more global state in the near term, but motion
actually feels and looks kind of good for once. There's also some work
left to do now related to the dig highlighting and collision response
based on the rendered camera height, but that will come as tweaks down
the road.