Consolidate mob data into Mob struct.
Similar reasoning to Player above.

For now, despite the data in Mob and Player being very similar, I'm
not going to introduce a hierarchy of structs or any intermediate
abstraction like Entity -- I'll just see how this copypasta goes
and add abstraction later, when I'm feeling the pain of a lack of it.