Add a working shadowmapped scene in new_main.cpp
Bring in the cavalry! This is a stripped down version of the scene I've
developed the past few weeks in the Learn OpenGL tutorials. It's using
c++ and adds GLM, GLFW, stb_image, GLAD, and some other libraries and
tools covered in the tutorial series.

This is a big of a mess, but it compiles (using `make new`) and runs
(using `./new_mc`), so it's a savepoint. Next up is some work to
simplify and clean it up, then to begin porting pieces of the existing
minecraft project in here -- we'll see if it all comes together!